test(go): consolidate *_branch_test.go into their sibling test files
Ludo's review asked to merge go/internal/utils/utils_branch_test.go
into utils_test.go. Apply the same to every branch-coverage file so the
package has one test file per source file rather than a split:

  utils_branch_test.go            -> utils_test.go
  charge_branch_test.go           -> charge_test.go
  transport_branch_test.go        -> transport_test.go
  server_branch_test.go           -> server_test.go
  server_more_branch_test.go      -> server_test.go

No test logic changes; imports reconciled with goimports.
